Revert "Revert "chore: routerをbulletproof-react参考にリファクタリング""
