+++ frontend/src/App.tsx
@@ -1,6 +1,8 @@
 import { FC } from 'react';
 // import './App.css';
 import { ChakraProvider } from '@chakra-ui/react';
+import { AccessTokenProvider } from 'hooks/providers/useAccessTokenProvider';
+import { AuthProvider } from 'hooks/providers/useAuthProvider';
 import { BrowserRouter } from 'react-router-dom';
 import { Router } from 'router/Router';
 import theme from './theme/theme';
@@ -10,7 +12,11 @@ const App: FC = () => {
     <div className="App">
       <ChakraProvider theme={theme}>
         <BrowserRouter>
-          <Router />
+          <AuthProvider>
+            <AccessTokenProvider>
+              <Router />
+            </AccessTokenProvider>
+          </AuthProvider>
         </BrowserRouter>
       </ChakraProvider>
     </div>
