+++ frontend/src/router/PublicRoutes.tsx
@@ -1,9 +1,8 @@
-import { FC } from 'react';
-import { Navigate, Outlet } from 'react-router-dom';
-import { useAuth } from '../hooks/providers/useAuthProvider';
+import { Login } from 'components/pages/Login';
 
-export const PublicRoutes: FC = () => {
-  const auth = useAuth();
-
-  return auth.user !== '' ? <Navigate to="/user-list" /> : <Outlet />;
-};
+export const publicRoutes = [
+  {
+    path: '/',
+    element: <Login />,
+  },
+];
